+++ themes/2006_v2/config.php
@@ -1,0 +1,8 @@
+<?php
+return [
+    'news_count' => 8,
+    'join_steam_text' => 'Join Steam for free and get games delivered straight to your desktop with automatic updates and a massive gaming community.',
+    'new_on_steam_header' => 'New On Steam',
+    'latest_news_header' => 'Latest News',
+    'catalog_header' => 'Find',
+];